# 🐛 Fix Windows support; ✨ Add options; ⚡ Improve performance