--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,12 +1,15 @@
 # Change Log
 
 ## v1.5.1
+
 - Added the `npm install` step to the build instructions in the readme
 
 ## v1.5.0
+
 - Transferred ownership to [maxcutlyp](https://github.com/maxcutlyp)
 
 ## v1.4.0
+
 - Migrated to TypeScript
 
 ## v1.3.0