--- a/README.md
+++ b/README.md
@@ -4,14 +4,10 @@ This extension provides auto-completion to any `process.env` mentions inside a J
 
 ## Install
 
-### From the Extension Marketplace
-
-Just search for `maxcutlyp.dotenv-autocomplete` in the extensions panel, and click install.
-
 ### From Source
 
 1. Install `vsce` with `npm install -g vsce`
-2. Clone this repo with `git clone https://github.com/maxcutlyp/dotenv-autocomplete`
+2. Clone this repo with `git clone https://github.com/younesaassila/dotenv-autocomplete`
 3. `cd` into the newly created folder and run `npm install` to install all the necessary packages for the extension to build
 4. Run `vsce package` - this will generate a file ending in .vsix
 5. Drag the generated file into the Extensions panel in VSCode (or run `code --install-extension <file-name>.vsix`)