--- a/package.json
+++ b/package.json
@@ -29,6 +29,9 @@
 		"pretest": "npm run compile && npm run lint",
 		"test": "node ./out/test/runTest.js"
 	},
+	"dependencies": {
+		"glob": "^8.0.3"
+	},
 	"devDependencies": {
 		"@types/glob": "^7.2.0",
 		"@types/mocha": "^9.1.0",
@@ -36,7 +39,6 @@
 		"@types/vscode": "^1.65.0",
 		"@vscode/test-electron": "^2.1.3",
 		"eslint": "^8.12.0",
-		"glob": "^7.2.0",
 		"mocha": "^9.2.2",
 		"typescript": "^4.6.3"
 	}