--- a/package.json
+++ b/package.json
@@ -34,12 +34,12 @@
 	},
 	"devDependencies": {
 		"@types/glob": "^7.2.0",
-		"@types/mocha": "^9.1.0",
-		"@types/node": "^17.0.23",
-		"@types/vscode": "^1.65.0",
-		"@vscode/test-electron": "^2.1.3",
-		"eslint": "^8.12.0",
+		"@types/mocha": "^9.1.1",
+		"@types/node": "^17.0.45",
+		"@types/vscode": "^1.70.0",
+		"@vscode/test-electron": "^2.1.5",
+		"eslint": "^8.21.0",
 		"mocha": "^9.2.2",
-		"typescript": "^4.6.3"
+		"typescript": "^4.7.4"
 	}
 }