--- a/package.json
+++ b/package.json
@@ -37,6 +37,8 @@
 		"@types/mocha": "^9.1.1",
 		"@types/node": "^17.0.45",
 		"@types/vscode": "^1.65.0",
+		"@typescript-eslint/eslint-plugin": "^5.33.0",
+		"@typescript-eslint/parser": "^5.33.0",
 		"@vscode/test-electron": "^2.1.5",
 		"eslint": "^8.21.0",
 		"mocha": "^9.2.2",