--- a/package.json
+++ b/package.json
@@ -62,6 +62,7 @@
 		"@typescript-eslint/parser": "^5.33.0",
 		"@vscode/test-electron": "^2.1.5",
 		"eslint": "^8.21.0",
+		"glob": "^8.0.3",
 		"mocha": "^9.2.2",
 		"typescript": "^4.7.4"
 	}