--- a/package.json
+++ b/package.json
@@ -51,6 +51,7 @@
 		"test": "node ./out/test/runTest.js"
 	},
 	"dependencies": {
+		"dotenv": "^16.0.1",
 		"fast-glob": "^3.2.11"
 	},
 	"devDependencies": {