--- a/package.json
+++ b/package.json
@@ -2,11 +2,11 @@
 	"name": "dotenv-autocomplete",
 	"displayName": "dotenv-autocomplete",
 	"description": "A VSCode extension that autocompletes process.env variables from .env files",
-	"version": "1.5.1",
-	"publisher": "maxcutlyp",
+	"version": "2.0.0",
+	"publisher": "younesaassila",
 	"repository": {
 		"type": "git",
-		"url": "https://github.com/maxcutlyp/dotenv-autocomplete"
+		"url": "https://github.com/younesaassila/dotenv-autocomplete"
 	},
 	"engines": {
 		"vscode": "^1.65.0"
@@ -51,20 +51,20 @@
 		"test": "node ./out/test/runTest.js"
 	},
 	"dependencies": {
-		"dotenv": "^16.0.1",
-		"fast-glob": "^3.2.11"
+		"dotenv": "^16.0.3",
+		"fast-glob": "^3.2.12"
 	},
 	"devDependencies": {
 		"@types/glob": "^7.2.0",
 		"@types/mocha": "^9.1.1",
 		"@types/node": "^17.0.45",
-		"@types/vscode": "^1.65.0",
-		"@typescript-eslint/eslint-plugin": "^5.33.0",
-		"@typescript-eslint/parser": "^5.33.0",
-		"@vscode/test-electron": "^2.1.5",
-		"eslint": "^8.21.0",
-		"glob": "^8.0.3",
+		"@types/vscode": "^1.77.0",
+		"@typescript-eslint/eslint-plugin": "^5.59.2",
+		"@typescript-eslint/parser": "^5.59.2",
+		"@vscode/test-electron": "^2.3.0",
+		"eslint": "^8.39.0",
+		"glob": "^8.1.0",
 		"mocha": "^9.2.2",
-		"typescript": "^4.7.4"
+		"typescript": "^4.9.5"
 	}
 }