--- a/package.json
+++ b/package.json
@@ -9,7 +9,7 @@
 		"url": "https://github.com/younesaassila/dotenv-autocomplete"
 	},
 	"engines": {
-		"vscode": "^1.65.0"
+		"vscode": "^1.77.0"
 	},
 	"categories": [
 		"Other"