Hello! I'm a user of your extension on Windows. (Thanks for your extension!)
For quite a while now, the extension hasn't worked properly on Windows.
This is why I forked your repo and set myself the goal of fixing it.
Here is what this pull request changes:

**Fixes**

- Project path was given to `glob` directly, which caused the Windows path separator and folders with special characters to make `glob` return an empty array when looking for `.env` files
- `findProjectDir` used to check for `dir = /` to end its logic, except that doesn't work on Windows because the root is `C:` (or whatever drive you're on)
- Improved performance by using `fast-glob` instead of `glob`. Did a quick benchmark:
  - Before: 1s 400ms
  - After: 0s 290ms
- Now using `dotenv`'s parser, thus:
  - Ignoring comments at end of line
- Added a newline at the end of files that didn't contain one at the end (to remove the ugly GitHub red sign at the end)
- Formatted the Markdown files with Prettier
- Upgraded dependencies
- Refactor `extension.ts`
- Update initial console log with new name of extension (`env-autocomplete` -> `dotenv-autocomplete`)

**New features**

- Added two config options to the extension:
  - `dotenv-autocomplete.envvarsScope`: Whether to suggest all environment variables or project-scoped ones only.
    - Possible values: `"all"` or `"project"`
    - This lets users hide environment variables like `PATH` or `TEMP` that are system specific and that may not get used.
  - `dotenv-autocomplete.showEnvvarsValues`: Show values of environment variables in the suggestions widget.
    - By default `false` because `.env` files usually contain secrets. This prevents leaking secrets when screen sharing or live streaming.

There's not a lot of changes in the code but a lot of commits caused I tried to do it cleanly by creating branches and creating commits for each separate change. Tested the code on both Windows and macOS.

**Not done, but to consider for the future:**

- Add an Issues tab to the GitHub repo
- Add tests